% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj_mat}
\alias{adj_mat}
\title{Adjacency matrix}
\usage{
adj_mat(mat, sym_rule = "and")
}
\arguments{
\item{mat}{matrix of regression coefficients}

\item{sym_rule}{symmetrization rule, either AND or OR}
}
\value{
adjacency matrix
}
\description{
Adjacency matrix
}
