% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_ols}
\alias{beta_ols}
\title{Initialize regression matrix}
\usage{
beta_ols(X)
}
\arguments{
\item{X}{data}
}
\value{
A zero-diagonal matrix of regression vectors.
}
\description{
Initialize regression matrix
}
