% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-cost.R, R/cost.R, R/utils.R
\name{beta_to_vector}
\alias{beta_to_vector}
\title{vectorize beta matrix}
\usage{
beta_to_vector(beta_mat)

beta_to_vector(beta_mat)

beta_to_vector(beta_mat)
}
\arguments{
\item{beta_mat}{matrix of regressions vectors}
}
\value{
A numeric vector of all regression coefficients.
}
\description{
vectorize beta matrix

vectorize beta matrix

Transform a matrix of regression coefficients to vector removing the diagonal
}
