% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-conesta_rwrapper.R
\name{conesta_rwrapper}
\alias{conesta_rwrapper}
\title{CONESTA solver for numerical experiments.}
\usage{
conesta_rwrapper(
  X,
  lam1,
  lam2,
  beta_warm = c(0),
  type_ = "initial",
  W_ = NULL,
  mean_ = FALSE,
  max_iter_ = 10000,
  prec_ = 0.01
)
}
\description{
CONESTA solver for numerical experiments.
}
\examples{
\dontrun{# because of installation of external packages during checks
mglasso::install_pylearn_parsimony(envname = "rmglasso", method = "conda")
reticulate::use_condaenv("rmglasso", required = TRUE)
reticulate::py_config()
n = 30
K = 2
p = 4
rho = 0.85
blocs <- list()
for (j in 1:K) {
 bloc <- matrix(rho, nrow = p/K, ncol = p/K)
   for(i in 1:(p/K)) { bloc[i,i] <- 1 }
   blocs[[j]] <- bloc
   }

mat.covariance <- Matrix::bdiag(blocs)
mat.covariance
set.seed(11)
X <- mvtnorm::rmvnorm(n, mean = rep(0,p), sigma = as.matrix(mat.covariance))
X <- scale(X)
res <- conesta_rwrapper(X, 0.1, 0.1)
}
}
