% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cost}
\alias{cost}
\title{`Mglasso` cost function}
\usage{
cost(beta, x, lambda1 = 0, lambda2 = 0)
}
\arguments{
\item{beta}{p by p numeric matrix. In rows, regression vectors coefficients after node-wise regression. \code{diag(beta) = 0}.}

\item{x}{n by p numeric matrix. Data with variables in columns.}

\item{lambda1}{numeric scalar. Lasso penalization parameter.}

\item{lambda2}{numeric scalar. Fused-group Lasso penalization parameter.}
}
\value{
numeric scalar. The cost.
}
\description{
`cost` computes the cost function of `Mglasso` method.
}
