% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-clustering.R, R/utils.R
\name{dist_beta}
\alias{dist_beta}
\title{distances Beta}
\usage{
dist_beta(beta, distance = "euclidean")

dist_beta(beta, distance = "euclidean")
}
\arguments{
\item{beta}{matrix of regression vectors}

\item{distance}{euclidean or relative distance}
}
\value{
A numeric matrix of distances.
}
\description{
distances Beta

Compute distance matrix between regression vectors
}
