% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-clustering.R
\name{expand_beta}
\alias{expand_beta}
\title{TO DO: Fill upper triangular matrix then sum up with the transpose to have full matrix}
\usage{
expand_beta(beta_level, clusters)
}
\description{
TO DO: Fill upper triangular matrix then sum up with the transpose to have full matrix
}
