% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-clustering.R
\name{expand_beta_deprecated}
\alias{expand_beta_deprecated}
\title{doesn't work when dealing with matrix where diagonal of zero should be adjusted}
\usage{
expand_beta_deprecated(beta_level, clusters)
}
\description{
doesn't work when dealing with matrix where diagonal of zero should be adjusted
}
