% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-perform.R
\name{get_auc}
\alias{get_auc}
\title{Plot ROC curve and calculate AUC}
\usage{
get_auc(omega_hat_list, omega, to = to_)
}
\arguments{
\item{type}{Classical ROC curve tpr = f(FPR) or TPR = f(precision) adjusted version.
Compute AUC and partial  AUC}
}
\description{
Plot ROC curve and calculate AUC
}
