% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-simulate.R
\name{get_range_nclusters}
\alias{get_range_nclusters}
\title{compute range of number of clusters from ROC outputs
take in parameter an object from reorder_mglasso_roc_calculations}
\usage{
get_range_nclusters(out, thresh_fuse = 1e-06, p = 40)
}
\description{
compute range of number of clusters from ROC outputs
take in parameter an object from reorder_mglasso_roc_calculations
}
