% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-plot.R
\name{ggplot_roc}
\alias{ggplot_roc}
\title{Title}
\usage{
ggplot_roc(
  omega_hat_list,
  omega,
  type = c("classical", "precision_recall"),
  main = NULL
)
}
\arguments{
\item{main}{}
}
\description{
Title
}
