% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{image_sparse}
\alias{image_sparse}
\title{Plot the image of a matrix}
\usage{
image_sparse(matrix, main_ = "", sub_ = "", col_names = FALSE)
}
\arguments{
\item{matrix}{matrix of regression coefficients}

\item{main_}{title}

\item{sub_}{subtitle}

\item{col_names}{columns names}
}
\description{
Plot the image of a matrix
}
