% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{install_libs}
\alias{install_libs}
\title{conesta_rwrapper
Install necessary libraries for CONESTA algorithm}
\usage{
install_libs()
}
\description{
conesta_rwrapper
Install necessary libraries for CONESTA algorithm
}
