% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-select_model_glasso.R
\name{lasso_estimate}
\alias{lasso_estimate}
\title{Check first estimate coeffs with glm}
\usage{
lasso_estimate(response_variable_number, penalty_value)
}
\description{
Check first estimate coeffs with glm
}
