% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-simulate.R
\name{mean_prec_mat}
\alias{mean_prec_mat}
\title{mean of randomly simulated precision matrices in the same configuration}
\usage{
mean_prec_mat(nrep = 10, config = config_)
}
\description{
mean of randomly simulated precision matrices in the same configuration
}
