% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-clustering.R
\name{merge_beta}
\alias{merge_beta}
\title{Merge Beta
Different types of merging and their effect}
\usage{
merge_beta(Beta, pair_to_merge, clusters)
}
\description{
Merge Beta
Different types of merging and their effect
}
