% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_clusters}
\alias{merge_clusters}
\title{compute clusters partition from pairs of variables to merge}
\usage{
merge_clusters(pairs_to_merge, clusters)
}
\arguments{
\item{pairs_to_merge}{table of the indices of variables to be merge}

\item{clusters}{numeric vector. By default 1:p where p is the number of variables}
}
\description{
compute clusters partition from pairs of variables to merge
}
