% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-clustering.R
\name{merge_proc}
\alias{merge_proc}
\title{merge clusters from table}
\usage{
merge_proc(
  pairs_to_merge,
  clusters,
  X,
  Beta,
  level,
  gain_level,
  gains,
  labels,
  merge
)
}
\description{
merge clusters from table
}
