% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglasso-package.R
\docType{package}
\name{mglasso-package}
\alias{mglasso}
\alias{mglasso-package}
\title{mglasso: Multiscale Graphical Lasso}
\description{
Inference of Multiscale graphical models with neighborhood selection approach. The method is based on solving a convex optimization problem combining a Lasso and fused-group Lasso penalties. This allows to infer simultaneously a conditional independence graph and a clustering partition. The optimization is based on the Continuation with Nesterov smoothing in a Shrinkage-Thresholding Algorithm solver (Hadj-Selem et al. 2018) \doi{10.1109/TMI.2018.2829802} implemented in python.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://desanou.github.io/mglasso/}
}

}
\author{
\strong{Maintainer}: Edmond Sanou \email{doedmond.sanou@univ-evry.fr}

Other contributors:
\itemize{
  \item Tung Le [contributor]
  \item Christophe Ambroise [thesis advisor]
  \item Geneviève Robin [thesis advisor]
}

}
\keyword{internal}
