% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_conesta.R
\name{mglasso_initialize}
\alias{mglasso_initialize}
\title{Initialize mglasso required python packages}
\usage{
mglasso_initialize(condaenv = "rmglasso")
}
\arguments{
\item{condaenv}{character. conda virtual environment name}
}
\description{
Initialize mglasso required python packages to call from R.
}
