% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-perform.R
\name{neighbor_select}
\alias{neighbor_select}
\title{neighbor_select}
\usage{
neighbor_select(
  data = data$X,
  config,
  lambda_min_ratio = 0.01,
  nlambda = 10,
  nresamples = 20,
  lambdas = NULL,
  model = NULL,
  verbose = FALSE,
  estim_var = NULL
)
}
\description{
neighbor_select
}
