% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-perform.R
\name{perf_one}
\alias{perf_one}
\title{compute TPR, FPR, SHD given estimated and true precision matrices}
\usage{
perf_one(omega_hat, omega)
}
\description{
compute TPR, FPR, SHD given estimated and true precision matrices
}
\details{
SHD: structural hamming distance
}
