% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mglasso}
\alias{plot.mglasso}
\title{Plot mglasso output}
\usage{
\method{plot}{mglasso}(mglasso_, beta_true_ = NULL, levels_ = NULL)
}
\arguments{
\item{mglasso_}{mglasso output}

\item{beta_true_}{provide true regression vectors if simulation model}

\item{levels_}{selected levels}
}
\description{
Plot mglasso output
}
