% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curve and calculate AUC}
\usage{
plot_roc(omega_hat_list, omega, type = "classical")
}
\arguments{
\item{type}{Classical ROC curve tpr = f(FPR) or TPR = f(precision) adjusted version.}

\item{to_}{maximum value for partial AUC. Default 1.}
}
\description{
Plot ROC curve and calculate AUC
}
\details{
When dealing with huge data sets with FPR and TPR do not vary at the same rate.
Low variation of number of edges has much more effect on TPR rather than FPR. It's better
to consider precision ie 1 - FDR (confidence) vs TPR (power) curve (type = "prec_rec").
}
