% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{precision_to_regression}
\alias{precision_to_regression}
\title{Compute precision matrix from regression vectors}
\usage{
precision_to_regression(K)
}
\arguments{
\item{K}{precision matrix}
}
\value{
A numeric matrix.
}
\description{
Compute precision matrix from regression vectors
}
