% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-select_model_glasso.R
\name{select_kfold}
\alias{select_kfold}
\title{K-fold cross validation neghborhood lasso selection}
\usage{
select_kfold(
  X,
  Thetas,
  lambdas = NULL,
  n_lambda,
  K_fold = 10,
  criterion = "ploglik",
  verbose = TRUE
)
}
\arguments{
\item{criterion}{c("ploglik", "rmse")}
}
\description{
K-fold cross validation neghborhood lasso selection
}
\details{
if criterion = "ploglik" use pseudo-log likelihood formula from Huge paper with matrix approach
if "rmse" use mean squarred prediction error
}
