% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-select_model_glasso.R
\name{select_partition}
\alias{select_partition}
\title{Finds the optimal number of clusters using slope heuristic}
\usage{
select_partition(gains)
}
\value{
integer scalar. The indice of the selected model.
}
\description{
Finds the optimal number of clusters using slope heuristic
}
