% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-select_model_mglasso.R
\name{select_stab_mglasso}
\alias{select_stab_mglasso}
\title{stability selection mglasso}
\usage{
select_stab_mglasso(X, l1_, l2_, subsample_ratio, nrep, stab_thresh)
}
\description{
stability selection mglasso
}
