% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-select_model_mglasso.R
\name{select_stars_mglasso}
\alias{select_stars_mglasso}
\title{stability selection mglasso II stars way}
\usage{
select_stars_mglasso(
  X,
  lambda1s = NULL,
  lambda2s = NULL,
  subsample_ratio = NULL,
  nrep = 1,
  stars_thresh = 0.1,
  nl1 = 1,
  nl2 = 1
)
}
\description{
stability selection mglasso II stars way
}
