% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-select_model_mglasso.R
\name{seq_l1l2}
\alias{seq_l1l2}
\title{def sequences for lambda1s and lambda2s
not sure if max of lambda1 s still the same as in the lasso case. But if find better equivalence will update this part}
\usage{
seq_l1l2(
  X,
  nlam1 = 2,
  nlam2 = 2,
  logscale = TRUE,
  mean = FALSE,
  lambda1_min_ratio = 0.01,
  require_non_list = FALSE,
  l2_max = NULL
)
}
\arguments{
\item{mean}{in conesta_rwrapper is the mean criterion used ie averaged by np}
}
\description{
def sequences for lambda1s and lambda2s
not sure if max of lambda1 s still the same as in the lasso case. But if find better equivalence will update this part
}
