% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_data}
\alias{sim_data}
\title{simulate data with given graph structure}
\usage{
sim_data(
  p = 20,
  np_ratio = 2,
  structure = c("from_tree", "scale_free", "stochastic_block", "planted",
    "block_diagonal"),
  sparsity = 0,
  alpha,
  prob_mat,
  members,
  K,
  rho,
  prop
)
}
\value{
A list:
tree_true: an object of class `phylo`. The true dendrogram, returned if the structure is simulated from a tree.
}
\description{
simulate data with given graph structure
}
