% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R, R/utils.R
\name{symmetrize}
\alias{symmetrize}
\title{symmetrize matrix of regression vectors pxp}
\usage{
symmetrize(mat, rule = "and")

symmetrize(mat, rule = "and")
}
\arguments{
\item{mat}{graph or precision matrix}

\item{rule}{"and" or "or" rule}
}
\value{
A numeric matrix.
}
\description{
symmetrize matrix of regression vectors pxp

Apply symmetrization on estimated graph
}
