% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-cost.R, R/cost.R, R/utils.R
\name{beta_ols}
\alias{beta_ols}
\title{Init Beta via OLS}
\usage{
beta_ols(X)

beta_ols(X)

beta_ols(X)
}
\arguments{
\item{X}{data}
}
\value{
A zero-diagonal matrix of regression vectors.
}
\description{
Init Beta via OLS

Init Beta via OLS

Initialize regression matrix
}
