% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_to_vector}
\alias{beta_to_vector}
\title{Transform a matrix of regression coefficients to vector removing the diagonal}
\usage{
beta_to_vector(beta_mat)
}
\arguments{
\item{beta_mat}{matrix of regressions vectors}
}
\value{
A numeric vector of all regression coefficients.
}
\description{
Transform a matrix of regression coefficients to vector removing the diagonal
}
