% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{dist_beta}
\alias{dist_beta}
\title{Compute distance matrix between regression vectors}
\usage{
dist_beta(beta, distance = "euclidean")
}
\arguments{
\item{beta}{matrix of regression vectors}

\item{distance}{euclidean or relative distance}
}
\description{
Compute distance matrix between regression vectors
}
