% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_lines}
\alias{fun_lines}
\title{weighted sum/difference of two regression vectors}
\usage{
fun_lines(i, j, beta, fun = `-`, ni = 1, nj = 1)
}
\arguments{
\item{i}{integer scalar. Indice of the first vector.}

\item{j}{integer scalar. Indice of the second vector.}

\item{beta}{p by p numeric matrix. In rows, regression vectors coefficients after node-wise regression. `diag(beta) = 0`.}

\item{fun}{function. Applied on lines.}

\item{ni}{integer scalar. Weight for vector `i`.}

\item{nj}{integer scalar. Wweight for vector `j`.}
}
\value{
numeric vector
}
\description{
`fun_lines` applies function `fun` to regression vectors while reordering the coefficients,
such that the `j`-th coefficient in `beta[j, ]` is permuted with the `i`-th coefficient.
}
\examples{
beta <- matrix(round(rnorm(9),2), ncol = 3)
diag(beta) <- 0
beta
fun_lines(1, 2, beta)
fun_lines(2, 1, beta)
}
