% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_conesta.R
\name{install_conesta}
\alias{install_conesta}
\title{Install the python library pylearn-parsimony and other required libraries}
\usage{
install_conesta(
  conda = "auto",
  extra_pack = c("scipy == 1.7.1", "scikit-learn", "numpy == 1.22.4", "six",
    "matplotlib"),
  conda_py_version = "3.8",
  restart_session = T
)
}
\arguments{
\item{conda}{Character. Path to conda executable. "auto" finds the path automatically.}

\item{extra_pack}{Character vector. Extra-packages to be installed.}

\item{conda_py_version}{Character. Python version. It is advised to use a version ">=3.7,<3.10".
The latest version under which mglasso was developped is scipy 1.7.1 which is based on python ">=3.7,<3.10".
In turn, this version of scipy can only be associated with a version of numpy ">=1.16.5,<1.23.0"}
}
\value{
No return value.
}
\description{
pylearn-parsimony contains the solver CONESTA used for the mglasso problem
and is available on github at https://github.com/neurospin/pylearn-parsimony
At this stage, we only support conda virtual environments.
}
