% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglasso.R
\name{merge_proc}
\alias{merge_proc}
\title{merge clusters from table}
\usage{
merge_proc(
  to_merge,
  clusters,
  x,
  beta,
  level,
  gain_level,
  gains,
  labels,
  merge
)
}
\description{
merge clusters from table
}
