% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglasso.R
\name{mglasso}
\alias{mglasso}
\title{Multiscale Graphical lasso}
\usage{
mglasso(
  x,
  lambda1 = 0,
  fuse_thresh = 10^-3,
  maxit = NULL,
  distance = "euclidean",
  lambda2_start = 1e-04,
  lambda2_factor = 1.5,
  precision = 0.01,
  weights_ = NULL,
  type = "initial",
  compact = TRUE
)
}
\arguments{
\item{x}{n by p numeric matrix. p-multivariate normal sample with n independent observations.}

\item{lambda1}{positive numeric scalar. Lasso penalty.}

\item{fuse_thresh}{positive numeric scalar. Threshold for cluster fusion.}

\item{maxit}{integer scalar. Maximum number of iterations.}

\item{distance}{character. Distance between regression vectors. Default euclidean;}

\item{lambda2_start}{numeric scalar. starting value for fused-group Lasso penalty (clustering penalty).}

\item{lambda2_factor}{numeric scalar. Step used to update fused-group Lasso penalty. `lambda2_factor*lambda2_start`.}

\item{precision}{precision of estimation algorithm.}

\item{weights_}{matrix of weights}

\item{type}{if "initial" use classical version of MGLasso without weights.}

\item{compact}{if TRUE, only save results when previous clusters are different from current}
}
\value{
A list.
\item{out}{list with matrix of regression vectors and clusters for each level.
Each element of the list contains two elements: `beta` and `clusters`}
\item{tree}{clustering tree}
}
\description{
Estimates a graphical model structure using Lasso and fused-group Lasso penalties.
Belongs to neighborhood selection inference methods.
}
\examples{
\dontrun{
n = 50
K = 3
p = 9
rho = 0.85
blocs <- list()
for (j in 1:K) {
  bloc <- matrix(rho, nrow = p/K, ncol = p/K)
  for(i in 1:(p/K)) { bloc[i,i] <- 1 }
  blocs[[j]] <- bloc
}

mat.covariance <- Matrix::bdiag(blocs)
mat.covariance

set.seed(11)
X <- mvtnorm::rmvnorm(n, mean = rep(0,p), sigma = as.matrix(mat.covariance))
X <- scale(X)

res <- mglasso(X, 0.1, lambda2_start = 0.1)
res$out[[1]]$clusters
res$out[[1]]$beta
}
}
