% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mglasso}
\alias{plot_mglasso}
\title{Plot \code{mglasso} function output.}
\usage{
plot_mglasso(mglasso_, levels_ = NULL)
}
\arguments{
\item{mglasso_}{Object of class \code{mglasso}.}

\item{levels_}{Character vector. Selected levels for which estimated matrices
will be plot. If NULL plot all levels.}
}
\value{
No return value.
}
\description{
Plot the object returned by the \code{mglasso} function.
}
