% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpt-simulate.R
\name{sim_data}
\alias{sim_data}
\title{simulate data with given graph structure}
\usage{
sim_data(
  p = 20,
  np_ratio = 2,
  structure = c("block_diagonal", "hub", "scale_free", "erdos"),
  alpha,
  prob_mat,
  rho,
  g,
  inter_cluster_edge_prob = 0.01,
  p_erdos = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{
A list:
graph : precision
}
\description{
simulate data with given graph structure
}
