% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symmetrize}
\alias{symmetrize}
\title{Apply symmetrization on estimated graph}
\usage{
symmetrize(mat, rule = "and")
}
\arguments{
\item{mat}{graph or precision matrix}

\item{rule}{"and" or "or" rule}
}
\value{
A numeric matrix.
}
\description{
Apply symmetrization on estimated graph
}
